Refactor: Move ObservableArray.ChangeListener one level up. Make it package scope.
